Update navbar: align items left, add credits and contact info
The `ul` element with the class `navbar-nav` has been modified to include the `me-auto` class, which aligns the navigation items to the left.

A new `span` element with the class `navbar-text` has been added to the navigation bar. This element provides credits to Melody14 and Annou, and includes a contact email (anas76@gmail.com) and an in-game contact (Annou).